Proyectos/Proyecto 2/docker/dnsapi/app/requirements.txt
flask
firebase-admin
dnslib
pyyaml